Remove sqlite cache size warning
The original performance problem on my machine doesn't seem to be
reproducable any more so maybe either sqlite got smarter or it was
something else entirely. Move the warning to DEBUG and soften the
documentation. If it still helps people can use it, but we don't have to
annoy them if it doesn't.

Signed-off-by: Florian Pritz <bluewind@xinu.at>