DB/remove_archive: Properly handle cases where the DB is empty after removal
Empty here means that the DB does not contain any backup archive
information because all information that is already in the db has to be
removed. The old code tries to copy existing data into the new table,
but since there are no archives to copy data for, the sql query fails.
Also, it would copy all rows and only populate the path column, but if
there are no archives that have timestamps, that's actually useless
work.

This patch ensures that the table is kept empty if there are no
archives.

Signed-off-by: Florian Pritz <bluewind@xinu.at>