Helper: Remove untaint_archive_name
We no longer need a special whitelist for archive names since the
database no longer uses them as column keys. We still need to untaint
variables that are passed to DBI so we use untaint() for this now. We
also move the location of the untaint call closer to its usage with
DBI/system() to prevent untainted data from leaking elsewhere.

Fixes #4

Signed-off-by: Florian Pritz <bluewind@xinu.at>