DB: Fix incorrect subpath handling for path that is a superstring of the previous path
Files that share a common substring (e.g. `/home/foo/.ssh/id_rsa` and
`/home/foo/.ssh/id_rsa.pub`) could incorrectly trigger an optimization
that was only supposed to be triggered if the first path is a directory.
The first path would then not be added to the database cache even though
it should have been.

Signed-off-by: Florian Pritz <bluewind@xinu.at>